+++ src/views/account/AccountViewContentMiddle.vue
@@ -8,20 +8,20 @@
     <v-col cols="12" md="6">
       <AccountViewCapitalization
         :capitalization="capital.capitalization"
-        :params="params"
+        :params="stakingParams"
       />
     </v-col>
     <v-col cols="12" md="6">
       <AccountViewDelegations
         :delegations="delegations"
-        :params="params"
+        :params="stakingParams"
         :validators="validators"
       />
     </v-col>
     <v-col cols="12" md="6">
       <AccountViewUnbondings
         :unbondings="unbondings"
-        :params="params"
+        :params="stakingParams"
         :validators="validators"
       />
     </v-col>
@@ -46,17 +46,14 @@ export default {
     AccountViewUnbondings,
   },
   computed: {
-    ...mapGetters('application', {
-      params: 'stakingParams',
-      validators: 'validators',
-    }),
-    ...mapGetters('account', {
-      balances: 'balances',
-      commission: 'commission',
-      delegations: 'delegations',
-      rewards: 'rewards',
-      unbondings: 'unbondings',
-    }),
+    ...mapGetters('application', ['stakingParams', 'validators']),
+    ...mapGetters('account', [
+      'balances',
+      'commission',
+      'delegations',
+      'rewards',
+      'unbondings',
+    ]),
     capital() {
       return accountBalanceHelper.build({
         balances: this.balances,
@@ -64,7 +61,7 @@ export default {
         delegations: this.delegations,
         rewards: this.rewards,
         unbondings: this.unbondings,
-        bondDenom: this.params.bond_denom,
+        bondDenom: this.stakingParams.bond_denom,
       });
     },
   },