+++ src/views/transactions/list/TransactionsViewListSearch.vue
@@ -8,7 +8,7 @@
               <v-autocomplete
                 :items="items"
                 :label="$t('labels.type')"
-                :loading="isLoading"
+                :loading="isLoadingTxs"
                 :placeholder="$t('msgs.startTypingToFilterTypes')"
                 :prepend-inner-icon="mdiMagnify"
                 dense
@@ -63,12 +63,9 @@ export default {
     selectedType: '',
   }),
   computed: {
-    ...mapGetters('transactions', {
-      isLoadingTxs: 'isLoading',
-      isAddingTxs: 'isAddingTxs',
-    }),
-    isLoading() {
-      return this.isLoadingTxs || this.isAddingTxs;
+    ...mapGetters('transactions', ['isLoading', 'isAddingTxs']),
+    isLoadingTxs() {
+      return this.isLoading || this.isAddingTxs;
     },
     items() {
       return TRANSACTIONS.SUPPORTED_TYPES;