+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Replace icons installation
+- Align style of actions and getters imports
 
 ## [3.0.11] - 2022-03-29